exec: extended windows workaround to handle non .\exe case
bin.is_relative() finds non-path prefixed commands ie "hello" is
non-path prefixed.  So drop case where file-name is the entire file.
EXCEPT - that means dropping the case where we @cd to a directory,
then run locally.

So replicate DOS behaviour manually and resolve to the exe if it
exists in the @cd dir.